server/pyproject.toml
[project]
name = "math-rag-assistant"
version = "3.1.0"
description = "Assistant mathématique avec RAG (Retrieval-Augmented Generation)"
authors = [
    { name = "Your Name", email = "your.email@example.com" },
]
requires-python = ">=3.12,<3.14"
readme = "README_REFACTORED.md"
license = { text = "MIT" }
keywords = ["math", "education", "rag", "ai", "langchain", "ollama"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Education",
    "Topic :: Education",
    "Topic :: Scientific/Engineering :: Mathematics",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

# Dépendances runtime
dependencies = [
    "chromadb>=1.3.0",
    "fastapi>=0.115.6",
    "langchain>=1.0.2",
    "langchain-core>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-chroma>=0.1.4",
    "langchain-ollama>=0.2.1",
    "langchain-text-splitters>=0.3.2",
    "ollama>=0.3.4",
    "pandas>=2.3.3",
    "pymupdf>=1.26.5",
    "pypdf>=6.1.3",
    "pyside6>=6.10.0",
    "pyside6-addons>=6.10.0",
    "python-dotenv>=1.2.1",
    "rank-bm25>=0.2.2",
    "rapidfuzz>=3.10.1",
    "rich>=14.2.0",
    "sse-starlette>=3.0.2",
    "uvicorn>=0.38.0",
    "sentence-transformers>=5.1.2",
]

[project.optional-dependencies]
# CLI avec Rich
cli = [
    "rich>=14.2.0",
]

# GUI PySide6
gui = [
    "pyside6>=6.10.0",
    "pyside6-addons>=6.10.0",
]

# Reranker pour améliorer le retrieval
reranker = [
    "sentence-transformers>=3.3.1",
    # Conseil: épingle toi-même torch (CPU/GPU) selon ta machine
    "torch>=2.5.1",
]

# Développement (linting = ruff)
dev = [
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "pytest-asyncio>=1.2.0",
    "black>=25.9.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.1",
    "ruff>=0.14.2",
]

# Documentation – par défaut Sphinx + Furo + MyST (recommandé)
docs-sphinx = [
    "Sphinx>=8.2.3",
    # "furo>=2025.9.25",
    "myst-parser>=4.0.1",
]

# (Optionnel) stack MkDocs si tu préfères le tout-Markdown
docs-mkdocs = [
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.6.22",
]

# Tout installer (runtime + extras utiles)
all = [
    # cli
    "rich>=14.2.0",
    # gui
    "pyside6>=6.10.0",
    "pyside6-addons>=6.10.0",
    # reranker
    "sentence-transformers>=3.3.1",
    "torch>=2.5.1",
    # dev
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "pytest-asyncio>=1.2.0",
    "black>=25.9.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.1",
    "ruff>=0.14.2",
    # docs sphinx + mkdocs
    "Sphinx>=8.2.3",
    "furo>=2025.9.25",
    "myst-parser>=4.0.1",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.6.22",
]

[project.urls]
Homepage = "https://github.com/yourusername/math-rag-assistant"
Documentation = "https://github.com/yourusername/math-rag-assistant/blob/main/README_REFACTORED.md"
Repository = "https://github.com/yourusername/math-rag-assistant"
Issues = "https://github.com/yourusername/math-rag-assistant/issues"

[project.scripts]
math-cli = "src.ui.cli.app:main"
math-server = "uvicorn:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/scripts",
    "/README_REFACTORED.md",
    "/README_MATH_RAG.md",
    "/MIGRATION.md",
    "/.env.example",
    "/pyproject.toml",
]

# =============================================================================
# Outils de développement
# =============================================================================

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312', 'py313']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
  | before
)/
'''

# Ruff remplace flake8 + isort (I rules)
[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "W", "RUF"]
ignore = ["E203", "W503"]

[tool.ruff.format]
# Laisse Black gérer le formatage si tu préfères (sinon: "preview = true" pour le formatter ruff)
preview = false

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "langchain.*",
    "langchain_ollama.*",
    "langchain_chroma.*",
    "langchain_community.*",
    "chromadb.*",
    "rich.*",
    "PySide6.*",
    "sentence_transformers.*",
    "fitz.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--strict-config",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/before/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

# =============================================================================
# UV (package manager moderne)
# =============================================================================

[tool.uv]
dev-dependencies = [
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "black>=25.9.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.1",
    "ruff>=0.14.2",
]

[tool.uv.sources]
# Si besoin de sources custom
# default = "pypi"
